support python3

+++ camera.py
@@ -11,11 +11,11 @@ __all__ = ['ProjectPoints3D', 'ProjectPoints', 'RigidTransform']
 
 import chumpy as ch
 from chumpy import depends_on, Ch
-from cvwrap import cv2
+from .cvwrap import cv2
 import numpy as np
 import scipy.sparse as sp
 from chumpy.utils import row, col
-from geometry import Rodrigues
+from .geometry import Rodrigues
 
 def RigidTransformSlow(**kwargs):
     # Returns a Ch object with dterms 'v', 'rt', and 't'
@@ -28,32 +28,32 @@ def RigidTransformSlow(**kwargs):
 
 class RigidTransform(Ch):
     dterms = 'v', 'rt', 't'
-    
+
     def compute_r(self):
         return (cv2.Rodrigues(self.rt.r)[0].dot(self.v.r.T) + col(self.t.r)).T.copy()
-        
+
     def compute_dr_wrt(self, wrt):
 
         if wrt not in (self.v, self.rt, self.t):
             return
-        
+
         if wrt is self.t:
-            if not hasattr(self, '_drt') or self._drt.shape[0] != self.v.r.size:                
+            if not hasattr(self, '_drt') or self._drt.shape[0] != self.v.r.size:
                 IS = np.arange(self.v.r.size)
                 JS = IS % 3
                 data = np.ones(len(IS))
                 self._drt = sp.csc_matrix((data, (IS, JS)))
             return self._drt
-        
+
         if wrt is self.rt:
             rot, rot_dr = cv2.Rodrigues(self.rt.r)
             rot_dr = rot_dr.reshape((3,3,3))
             dr = np.einsum('abc, zc -> zba', rot_dr, self.v.r).reshape((-1,3))
             return dr
-        
+
         if wrt is self.v:
             rot = cv2.Rodrigues(self.rt.r)[0]
-            
+
             IS = np.repeat(np.arange(self.v.r.size), 3)
             JS = np.repeat(np.arange(self.v.r.size).reshape((-1,3)), 3, axis=0)
             data = np.vstack([rot for i in range(self.v.r.size/3)])
@@ -68,14 +68,14 @@ class ProjectPoints(Ch):
     def is_valid(self):
         if any([len(v.r.shape) > 1 for v in [self.rt, self.t, self.f, self.c, self.k]]):
             return False, 'rt, t, f, c, and k must be 1D'
-        
+
         if any([v.r.size != 3 for v in [self.rt, self.t]]):
             return False, 'rt and t must have size=3'
-        
+
         if any([v.r.size != 2 for v in [self.f, self.c]]):
             return False, 'f and c must have size=2'
 
-        return True, '' 
+        return True, ''
 
     def compute_r(self):
         return self.r_and_derivatives[0].squeeze()
@@ -84,7 +84,7 @@ class ProjectPoints(Ch):
     def compute_dr_wrt(self, wrt):
         if wrt not in [self.v, self.rt, self.t, self.f, self.c, self.k]:
             return None
-            
+
         j = self.r_and_derivatives[1]
         if wrt is self.rt:
             return j[:, :3]
@@ -99,8 +99,8 @@ class ProjectPoints(Ch):
         elif wrt is self.v:
             rot = cv2.Rodrigues(self.rt.r)[0]
             data = np.asarray(j[:, 3:6].dot(rot), order='C').ravel()
-            IS = np.repeat(np.arange(self.v.r.size*2/3), 3)
-            JS = np.asarray(np.repeat(np.arange(self.v.r.size).reshape((-1,3)), 2, axis=0), order='C').ravel()
+            IS = np.repeat(np.arange(self.v.r.size*2/3), 3).astype(int)
+            JS = np.asarray(np.repeat(np.arange(self.v.r.size).reshape((-1,3)), 2, axis=0), order='C').ravel().astype(int)
             result = sp.csc_matrix((data, (IS, JS)))
             return result
 
@@ -129,7 +129,7 @@ class ProjectPoints(Ch):
         xyz = self.unproject_points(uvd, camera_space=camera_space)
         return xyz.reshape((depth_image.shape[0], depth_image.shape[1], -1))
 
-    
+
     @depends_on('f','c')
     def camera_mtx(self):
         return np.array([[self.f.r[0], 0, self.c.r[0]],[0., self.f.r[1], self.c.r[1]],[0.,0.,1.]], dtype=np.float64)
@@ -138,30 +138,30 @@ class ProjectPoints(Ch):
     def view_mtx(self):
         R = cv2.Rodrigues(self.rt.r)[0]
         return np.hstack((R,col(self.t.r)))
-        
+
     @depends_on('v', 'rt', 't', 'f', 'c', 'k')
     def r_and_derivatives(self):
         v = self.v.r.reshape((-1,3)).copy()
         return cv2.projectPoints(v, self.rt.r, self.t.r, self.camera_mtx, self.k.r)
-        
+
     @property
     def view_matrix(self):
         R = cv2.Rodrigues(self.rt.r)[0]
         return np.hstack((R, col(self.t.r)))
 
-      
+
 class ProjectPoints3D(ProjectPoints):
     dterms = 'v', 'rt', 't', 'f', 'c', 'k'
 
     def compute_r(self):
         result = ProjectPoints.compute_r(self)
         return np.hstack((result, col(self.z_coords.r)))
-        
+
     @property
     def z_coords(self):
         assert(self.v.r.shape[1]==3)
         return RigidTransform(v=self.v, rt=self.rt, t=self.t)[:,2]
-    
+
     def compute_dr_wrt(self, wrt):
         result = ProjectPoints.compute_dr_wrt(self, wrt)
         if result is None:
@@ -171,11 +171,11 @@ class ProjectPoints3D(ProjectPoints):
             drz = self.z_coords.dr_wrt(wrt).tocoo()
             result = result.tocoo()
             result.row = result.row*3/2
-            
+
             IS = np.concatenate((result.row, drz.row*3+2))
             JS = np.concatenate((result.col, drz.col))
             data = np.concatenate((result.data, drz.data))
-            
+
             result = sp.csc_matrix((data, (IS, JS)), shape=(self.v.r.size, wrt.r.size))
         else:
             bigger = np.zeros((result.shape[0]/2, 3, result.shape[1]))
@@ -188,14 +188,14 @@ class ProjectPoints3D(ProjectPoints):
 
             result = bigger.reshape((-1, bigger.shape[-1]))
 
-        return result            
-            
-      
+        return result
+
+
 
 def main():
-    
+
     import unittest
-    from test_camera import TestCamera
+    from .test_camera import TestCamera
     suite = unittest.TestLoader().loadTestsFromTestCase(TestCamera)
     unittest.TextTestRunner(verbosity=2).run(suite)
 
@@ -203,4 +203,3 @@ def main():
 
 if __name__ == '__main__':
     main()
-
